+++ terraform/variables.tf
@@ -177,7 +177,7 @@ variable "num_vault_pods" {
 
 variable "vault_container" {
   type        = string
-  default     = "vault:1.1.3"
+  default     = "vault:1.2.0"
   description = "Name of the Vault container image to deploy. This can be specified like \"container:version\" or as a full container URL."
 }